requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
elevenlabs==1.12.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
obs-websocket-py==1.0
playsound==1.3.0
pycairo==1.27.0
pydantic==2.9.2
pydantic_core==2.23.4
PyGObject==3.50.0
python-dotenv==1.0.1
python-telegram-bot==21.7
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
